Added structure and most functions
I need to write the whole iteration function. It's not been tested past that, and I may try iteration with a placeholder now. I am getting some weird error in repeatPrompt()'s print statement but we'll see. The hardest parts should be board generation and iterating and if I've got the first down I'm halfway there.